Flatten Note
